Add gRPC client, rules of discount and refactor productId from long to int

+++ src/main/java/com/victorparanhos/ecommerceservice/applicationcore/domain/commands/CheckoutItemCommand.java
@@ -1,15 +1,15 @@
 package com.victorparanhos.ecommerceservice.applicationcore.domain.commands;
 
 public class CheckoutItemCommand {
-    private final long productId;
+    private final int productId;
     private final long quantity;
 
-    public CheckoutItemCommand(long id, long quantity) {
+    public CheckoutItemCommand(int id, long quantity) {
         this.productId = id;
         this.quantity = quantity;
     }
 
-    public long getProductId() {
+    public int getProductId() {
         return productId;
     }
